+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 - doc comments
 
+### Changed
+
+- return byte array instead of `heapless::String` in public API
+
 ## [0.1.0] - 2026-08-14
 
 ### Added